requirements.txt
Django>=5.2,<5.3
djangorestframework==3.16.1
django-filter==25.2
psycopg2-binary==2.9.10